Moved repository test config to separate class.
